factory/requirements.txt
asyncssh>=2.13.1
docker>=6.1.2
fastapi>=0.95.1
paramiko>=3.1.0
pyyaml>=6.0
types-PyYAML>=6.0.12.9
uvicorn>=0.22.0
